refactor(hw_41): optimization files v_2
