+++ react/my-app/src/components/Reg.tsx
@@ -1,6 +1,6 @@
 import { useContext, useState } from "react"
-import "../styles/reg.css"
 import { ThemeContext } from "../tools/themeContext"
+import "../styles/reg.css"
 
 type RegProps = {
     onReg: (login: string, password: string) => void
