+++ react/my-app/src/components/Search.tsx
@@ -1,13 +1,14 @@
+import { useState } from "react";
 import "../styles/search.css";
-
-type SearchProps = { value: string }
 
 export const Search = () => {
 
+    const [searchValue, setSearchValue] = useState("")
+
     return (
         <div className="header-search">
-            <input className="header-search__input" type="text" placeholder="search" />
-            <button className="header-search__button"></button>
+            <input className="header-search__input" type="search" placeholder="search" value={searchValue} onChange={(e) => setSearchValue(e.target.value)} />
+            <button className="header-search__button" onClick={() => searchValue}></button>
         </div>
     )
 }
