+++ react/my-app/src/tools/getPosts.ts
@@ -21,7 +21,7 @@ type PostsResponse = {
 
 type PostsParams = { limit?: number };
 
-export const getPosts = async ({ limit }: PostsParams) => {
+export const getPosts = async ({ limit}: PostsParams) => {
   const postsUrl = new URL(DOMAIN + POSTS);
   if (limit) postsUrl.searchParams.set("limit", String(limit));
   const response = await fetch(postsUrl);
